move incidents/alerts tests to incidents_alerts_tests folder
